Test and extend the MonetDialect for all Spark Types
Integer spark types now go to a larger sql type because of
the NULL representation.

Also because Postgres does it, but not always.

Reading interval types works, writing is not supported.
I think this is due to Spark itself but I'm not sure.
You get an error 'type not supported by DataSource' that seems
hardcoded.