Refactor data handling
Fixes #3

Before we had Conversions which each did all the work of extracting a
value, converting it and storing it in the Collector.  This was mostly ok
when all data types were mapped to the same type, but now our Dialect
sometimes maps values to larger types so the number of combinations just
becomes too big.

Instead of Conversions we now have Steps. Extracting, converting,
checking and storing are now distinct steps so we can mix and match.